Bump Pipeline, k8s, and Knative dependencies
This change uses non-tagged commits from `knative.dev/serving` and `knative.dev/eventing` because
their `v0.31.0` releases require a minimum k8s cluster version of 1.22, and I wanted to avoid
requiring that if at all possible. Instead, we depend on the latest commit in each before they
changed to depending on a `knative.dev/pkg` commit newer than the one Pipeline depends on.

Also add a Makefile, sync `.golangci.yml` with Pipeline's (other than requiring comments for
exported functions/types/etc, which I'll come back and do in a followup), and fix some new-as-a-result
linter issues.

+++ cmd/controller/main.go
@@ -46,15 +46,15 @@ var (
 	writeTimeOut          = flag.Int64("el-writetimeout", elresources.DefaultWriteTimeout, "The write timeout for EventListener Server.")
 	idleTimeOut           = flag.Int64("el-idletimeout", elresources.DefaultIdleTimeout, "The idle timeout for EventListener Server.")
 	timeOutHandler        = flag.Int64("el-timeouthandler", elresources.DefaultTimeOutHandler, "The timeout for Timeout Handler of EventListener Server.")
-	httpClientReadTimeOut = flag.Int64("el-httpclient-readtimeout", int64(elresources.DefaultHTTPClientReadTimeOut),
+	httpClientReadTimeOut = flag.Int64("el-httpclient-readtimeout", elresources.DefaultHTTPClientReadTimeOut,
 		"The HTTP Client read timeout for EventListener Server.")
-	httpClientKeepAlive = flag.Int64("el-httpclient-keep-alive", int64(elresources.DefaultHTTPClientKeepAlive),
+	httpClientKeepAlive = flag.Int64("el-httpclient-keep-alive", elresources.DefaultHTTPClientKeepAlive,
 		"The HTTP Client read timeout for EventListener Server.")
-	httpClientTLSHandshakeTimeout = flag.Int64("el-httpclient-tlshandshaketimeout", int64(elresources.DefaultHTTPClientTLSHandshakeTimeout),
+	httpClientTLSHandshakeTimeout = flag.Int64("el-httpclient-tlshandshaketimeout", elresources.DefaultHTTPClientTLSHandshakeTimeout,
 		"The HTTP Client read timeout for EventListener Server.")
-	httpClientResponseHeaderTimeout = flag.Int64("el-httpclient-responseheadertimeout", int64(elresources.DefaultHTTPClientResponseHeaderTimeout),
+	httpClientResponseHeaderTimeout = flag.Int64("el-httpclient-responseheadertimeout", elresources.DefaultHTTPClientResponseHeaderTimeout,
 		"The HTTP Client read timeout for EventListener Server.")
-	httpClientExpectContinueTimeout = flag.Int64("el-httpclient-expectcontinuetimeout", int64(elresources.DefaultHTTPClientExpectContinueTimeout),
+	httpClientExpectContinueTimeout = flag.Int64("el-httpclient-expectcontinuetimeout", elresources.DefaultHTTPClientExpectContinueTimeout,
 		"The HTTP Client read timeout for EventListener Server.")
 	periodSeconds    = flag.Int("period-seconds", elresources.DefaultPeriodSeconds, "The Period Seconds for the EventListener Liveness and Readiness Probes.")
 	failureThreshold = flag.Int("failure-threshold", elresources.DefaultFailureThreshold, "The Failure Threshold for the EventListener Liveness and Readiness Probes.")
